pyproject.toml
[project]
name = "cis7000-aurora"
version = "0.1.0"
description = "Aurora atmospheric foundation model — inference and fine-tuning on GCP"
requires-python = ">=3.10"

dependencies = [
    "microsoft-aurora",
    "torch",
    "google-cloud-storage",
    "numpy",
    "xarray",
    "netCDF4",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "ipykernel",
    "pytest",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
